commit intégration sprint2

+++ back_matagri_projet_3/src/main/java/mg/inclusiv/cdan8/projet3/Entities/Materiels.java
@@ -30,10 +30,6 @@ public class Materiels {
     private String techniqueMat;
     @Column(nullable = false) 
     private String imagePath;
-
-    // @ManyToOne
-    // @JoinColumn(name = "id_user")
-    // private Users user;
-    private Integer id_user;
+    private Long id_user;
     
 }
